Add rudimentary multiprocessing to QCD density.
Use Python's multiprocessing module in the pdf() method to start 4
worker processes in parallel. Each worker gets an equal chunk of the
data to work on. This gives gives a huge speedup in batch processing.

The method could still be improved. The workers actually need to be
initialized only once and can be kept running. Each time pdf() is
called, data can be passed to the workers.